feat: implement ConfigurationsViewModel for improved state management and business logic in settings screen
